Implement the `Default` trait for Settings struct
We had already implemented a `Settings::default()` method but recently
found out that there is a `Default` trait which does pretty much the
same thing but auto-resolves to the correct type. This commit removes
the `Settings::default()` method and moves the logic to `Default`'s
implementation for the `Settings` struct.

Additionally, makes the settings parsing code a bit easier to read by
removing some `match`es, replacing them with `unwrap_or` for the default
value.

Lastly, paves the way for a possible change that will allow worker
numbers to be non-positive