Rename settings module to config
This commit renames the settings module to config (as well as the
`Settings` struct to `Config`). Additionally, changes a bit the way
settings are retrieved so instead of top-level methods with in-name
namespacing (e.g. `db_user()`), there are now top-level methods
returning the specific config object, which exposes its fields as
functions (e.g. `cfg.db().user()`)