Bump yara crate to v0.21.0
The new version changes the crate's API. This commit includes the
required changes to make it compilable too.

Additionally (and more importantly), the new version provides a
`vendored` feature. When used, the Yara engine is also compiled during
dependency resolution.
Ref: https://github.com/Hugal31/yara-rust/blob/master/yara-sys/README.md

Due to the above, the extra setup steps for installing yara have been
removed, both from the Dockerfile, as well as the README