Introduce Feeder
The function Feeder.start_feeders spawns N threads, each of which
immediately blocks on a Redis topic, waiting to pop the first available
event. When an event is popped, it is deserialized into an Event object
and written into a crossbeam channel, on the read-end of which exist
processor threads. The number of feeder threads to be spawned is
configurable through `workers:feeders` in `config.yaml`

Ref #6